backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
google-generativeai
pytest
pytest-asyncio
httpx
